cold storage improvements
* pretty sure this fixes #81 and it definitley fixes #71
* the various generic secret types now support tags on freeze/thaw
* assorted refactoring/bug fixes
