+++ aomi/helpers.py
@@ -209,3 +209,17 @@ def actually_flatten(iterable):
             remainder = IT.chain(first, remainder)
         else:
             yield first
+
+
+def subdir_path(directory, relative):
+    """Returns a file path relative to another path."""
+    item_bits = directory.split(os.sep)
+    relative_bits = relative.split(os.sep)
+    for i in range(0, len(item_bits)):
+        if i == len(relative_bits) - 1:
+            return os.sep.join(item_bits[i:])
+        else:
+            if item_bits[i] != relative_bits[i]:
+                return None
+
+    return None
