+++ tests/integration/fixtures/cold/Secretfile
@@ -7,6 +7,13 @@ secrets:
     name: 'secret'
   mount: 'secret'
   path: 'foo'
+- files:
+  - source: 'sub/secret.txt'
+    name: 'secret'
+  mount: 'secret'
+  path: 'sub'
+  tags:
+  - sub
 - var_file: 'secret.yml'
   mount: 'secret'
   path: 'bar'
